+++ LaboraGTv0.1/LaboraGTv0.1/Models/EstadoCivil.cs
@@ -20,7 +20,7 @@ namespace LaboraGTv0._1.Models
             this.Colaborador = new HashSet<Colaborador>();
         }
     
-        public decimal idEstadoCivil { get; set; }
+        public int idEstadoCivil { get; set; }
         public string descripcion { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
