+++ LaboraGTv0.1/LaboraGTv0.1/Models/NivelAcademico.cs
@@ -20,7 +20,7 @@ namespace LaboraGTv0._1.Models
             this.DatosAcademicos = new HashSet<DatosAcademicos>();
         }
     
-        public decimal idNivelAcademico { get; set; }
+        public int idNivelAcademico { get; set; }
         public string descripcion { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
